safety_management_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
qrcode[pil]==7.3.1
requests==2.31.0